exampleEdited.c
This runs the klt feature tracker on the example images and prints out the features for each frame.  Basically I had to go through and renumber the features so we could use the information for the next step of our project.  This information is stored in a 3d array which corresponds the the info for each feature in each frame (the info is the feature number, x coordinate, y coordinate, and value).  I still need output it to a file to actually make it usable but I figured put this up now as an update of what I have been working on.