simple_bundle_adjuster.cc
So, if you scroll down to the end you will see the code I added. It now prints out the first 100 points (x,y, z coordinates) before and after the bundle adjuster improves the guess.  I was doing this to get an understand of how the start-out-of-the-package bundle adjuster worked and to understand how to retrieve the output (the information we actually want to reconstruct the scene!). Let me know what you think.